HW8 [Realization of meals with datatables] (without filtering and edit): - refactor already added MealUIController - from meals.jsp comes x-www-form-urlencoded (@RequestBody (JSON) changed to @RequestParam's); - "Are you sure?" check for delete operations move to deleteRow function (topjava.commons,js), call of this function attach to the onclick event if meals.jsp and users.jsp; - mealAjaxUrl changed from "/profile/meals/" to "profile/meals/" (404 response status fix).
NOTE: after adding new user or meal - jsp formatting dont work.

+++ src/main/webapp/resources/js/topjava.common.js
@@ -4,11 +4,6 @@ function makeEditable(datatableApi) {
     ctx.datatableApi = datatableApi;
 
     form = $('#detailsForm');
-    $(".delete").click(function () {
-        if (confirm('Are you sure?')) {
-            deleteRow($(this).closest('tr').attr("id"));
-        }
-    });
 
     $(document).ajaxError(function (event, jqXHR, options, jsExc) {
         failNoty(jqXHR);
@@ -24,13 +19,15 @@ function add() {
 }
 
 function deleteRow(id) {
-    $.ajax({
-        url: ctx.ajaxUrl + id,
-        type: "DELETE"
-    }).done(function () {
-        updateTable();
-        successNoty("Deleted");
-    });
+    if (confirm("Are you sure?")) {
+        $.ajax({
+            url: ctx.ajaxUrl + id,
+            type: "DELETE"
+        }).done(function () {
+            updateTable();
+            successNoty("Deleted");
+        });
+    }
 }
 
 function updateTable() {
@@ -43,8 +40,7 @@ function save() {
     const form = $("#detailsForm");
     $.ajax({
         type: "POST",
-        url: "http://localhost:8080/topjava/profile/meals",
-        contentType: "application/json;charset=UTF-8",
+        url: ctx.ajaxUrl,
         data: form.serialize()
     }).done(function () {
         $("#editRow").modal("hide");
